Update folder and code structures
The new folder configuration is divided in "db", "results", "src" and main code. The "results" division has "primary", "quantitative", "graph" and "visualization" (html) as directories. The code was divided in "main", which collects data from data base, and "src/counter", which uses the primary files to arrive at numerical results.