Add general analysis, new database and fields of research
The general analysis was described at the README.md file. The new database is the "pq" folder, which contains the fifteen thousand most productive researchers in the country. The "results" folder was updated with the new files about "pq" database. The mining code was updated to collect the fields of research of each curriculum and the counter code was updated to count the number of researchers in each field of research. All code was cleaned and commented